SERVER-17414 Disable and fix warnings
GitOrigin-RevId: 777df4cb365edf0d7227f0583d52e40c87dbd872

+++ src/mongo/db/modules/enterprise/src/ldap/connections/windows_ldap_connection.cpp
@@ -221,7 +221,7 @@ Status WindowsLDAPConnection::bindAsUser(const LDAPBindOptions& options) {
 }
 
 StatusWith<LDAPEntityCollection> WindowsLDAPConnection::query(LDAPQuery query) {
-    l_timeval timeout{_timeoutSeconds, 0};
+    l_timeval timeout{static_cast<LONG>(_timeoutSeconds), 0};
     return _pimpl->query(std::move(query), &timeout);
 }
